fix react dom error
