+++ components/blocks/features.tsx
@@ -4,9 +4,10 @@ import {
   PageBlocksFeaturesItems,
 } from "../../tina/__generated__/types";
 import { tinaField } from "tinacms/dist/react";
-import { Icon, iconSchema } from "../icon";
+import { Icon } from "../icon";
 import { Section } from "../layout/section";
 import { Container } from "../layout/container";
+import { iconSchema } from "../../tina/fields/icon";
 
 export const Feature = ({
   featuresColor,
