+++ package.json
@@ -25,20 +25,25 @@
     "postcss-nesting": "^12.1.5"
   },
   "dependencies": {
-    "@headlessui/react": "^2.1.0",
+    "@headlessui/react": "^2.1.1",
     "@tailwindcss/typography": "^0.5.8",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.1",
     "date-fns": "^3.6.0",
     "next": "^14.2.4",
     "next-themes": "^0.3.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.2.1",
     "styled-jsx": "^5.1.1",
     "tailwind-merge": "^2.3.0",
     "tailwindcss": "^3.2.4",
     "tinacms": "^1.6.4",
     "typescript": "^5.4.5"
+  },
+  "resolutions": {
+    "_#IMPORTANT": "The below resolutions are required due to tinacms being on react 17",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   }
 }
